+++ docs/rules/cred-class-field-unobscured-sensitive-input.md
@@ -1,6 +1,6 @@
 [//]: # "File generated from a template. Do not edit this file directly."
 
-# cred-class-field-type-options-missing-from-sensitive-input
+# cred-class-field-unobscured-sensitive-input
 
 `typeOptions.password` must be set to `true` in a sensitive node parameter, to obscure the input.
 
@@ -81,5 +81,5 @@ export class TestApi implements ICredentialType {
 
 ## Links
 
-- [Rule source](../../lib/rules/cred-class-field-type-options-missing-from-sensitive-input.ts)
-- [Test source](../../tests/cred-class-field-type-options-missing-from-sensitive-input.test.ts)
+- [Rule source](../../lib/rules/cred-class-field-unobscured-sensitive-input.ts)
+- [Test source](../../tests/cred-class-field-unobscured-sensitive-input.test.ts)
