+++ lib/rules/cred-class-field-unobscured-sensitive-input.ts
